move to mmocr/core
