+++ mmocr/core/deployment/__init__.py
@@ -1,0 +1,11 @@
+from .deploy_helper import (ONNXRuntimeRecognizer,
+                            ONNXRuntimeDetector,
+                            TensorRTDetector,
+                            TensorRTRecognizer)
+
+__all__ = [
+    'ONNXRuntimeRecognizer',
+    'ONNXRuntimeDetector',
+    'TensorRTDetector',
+    'TensorRTRecognizer'
+]
