+++ tools/deployment/onnx2tensorrt.py
@@ -12,10 +12,9 @@ from mmcv.tensorrt import is_tensorrt_plugin_loaded, onnx2trt, save_trt_engine
 from mmdet.datasets import replace_ImageToTensor
 from mmdet.datasets.pipelines import Compose
 
+from mmocr.core.deployment import (ONNXRuntimeDetector, ONNXRuntimeRecognizer,
+                                   TensorRTDetector, TensorRTRecognizer)
 from mmocr.datasets.pipelines.crop import crop_img  # noqa: F401
-from mmocr.models.deploy_helper import (ONNXRuntimeDetector,
-                                        ONNXRuntimeRecognizer,
-                                        TensorRTDetector, TensorRTRecognizer)
 
 
 def get_GiB(x: int):
