+++ tools/deployment/pytorch2onnx.py
@@ -11,9 +11,8 @@ from mmdet.datasets import replace_ImageToTensor
 from mmdet.datasets.pipelines import Compose
 from torch import nn
 
+from mmocr.core.deployment import ONNXRuntimeDetector, ONNXRuntimeRecognizer
 from mmocr.datasets.pipelines.crop import crop_img  # noqa: F401
-from mmocr.models.deploy_helper import (ONNXRuntimeDetector,
-                                        ONNXRuntimeRecognizer)
 
 
 def _convert_batchnorm(module):
